\name{DepsCheckeR-package}
\alias{DepsCheckeR-package}
\alias{DepsCheckeR}
\docType{package}
\title{CRAN dependencies analysis tool}
\description{
  This package contains functions to analyse CRAN packages dependencies
  using both information in packages DESCRIPTION files and results of R
  CMD check available on CRAN.
}
\details{
\tabular{ll}{
Package: \tab DepsCheckeR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-21\cr
License: \tab GPL-3\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
