\name{DepsCheckeR-package}
\alias{DepsCheckeR-package}
\alias{DepsCheckeR}
\docType{package}
\encoding{UTF-8}
\title{CRAN dependencies analysis tool}
\description{
  This package contains functions to analyse CRAN packages dependencies
  using both information in packages DESCRIPTION files and results of R
  CMD check available on CRAN.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\keyword{dependencies}
\keyword{cran}
\keyword{checking}
\keyword{R CMD check}